Fix: Upadated Navbar UI and navbar commented on Index page
