+++ src/pages/index.tsx
@@ -9,6 +9,7 @@ import AddPosts from "~/components/AddPosts";
 import React from "react";
 import Sidebar from "~/components/SideBar";
 import PostView from "~/components/Post";
+import Navbar from "~/components/Navbar";
 
 const Home: NextPage = () => {
   const { data, isLoading } = api.posts.getAll.useQuery();
@@ -20,29 +21,29 @@ const Home: NextPage = () => {
       <Head>
         <title>GreenGift</title>
         <meta name="description" content="Generated by create-t3-app" />
-        <link rel="icon" href="/favicon.ico" />
+        <link rel="icon" href="/Green-gift-favicon.png" />
       </Head>
-      <main className="flex">
-        <div className="flex w-full">
-          <div className="w-1/6">
+      <main>
+        {/* <div> */}
+        {/* <div className="w-1/6">
             <Sidebar />
+          </div> */}
+        <div className="">
+          <div>
+            <Navbar />
           </div>
-          <div className="w-4/6">
-            <div className="flex ">
-              <Navbar />
-            </div>
-            {/* <div className="flex"> */}
-            <div className="grid grid-cols-3 gap-4">
-              {data?.map((post) => (
-                <div key={post.id} className="rounded-lg bg-white shadow-md">
-                  <PostView postId={post.id} post={post} />
-                </div>
-              ))}
-            </div>
-            {/* </div> */}
+          {/* <div className="flex"> */}
+          <div className="mx-10 grid grid-cols-3 gap-4">
+            {data?.map((post) => (
+              <div key={post.id} className="rounded-lg bg-white shadow-md">
+                <PostView postId={post.id} post={post} />
+              </div>
+            ))}
           </div>
-          <div className="w-1/6"></div>
+          {/* </div> */}
         </div>
+        {/* <div className="w-1/6"></div> */}
+        {/* </div> */}
       </main>
     </>
   );
@@ -50,56 +51,64 @@ const Home: NextPage = () => {
 
 export default Home;
 
-const Navbar: React.FC = () => {
-  const { data: sessionData } = useSession();
-  const [showModal, setShowModal] = React.useState(false);
-  const openModal = () => setShowModal(true);
-  const closeModal = () => setShowModal(false);
+// const Navbar: React.FC = () => {
+//   const { data: sessionData } = useSession();
+//   const [showModal, setShowModal] = React.useState(false);
+//   const openModal = () => setShowModal(true);
+//   const closeModal = () => setShowModal(false);
 
-  return (
-    <div className="w-full border-slate-400">
-      <div className="flex gap-5 border-b border-slate-400 p-4">
-        <div className=" flex w-full gap-3">
-          {sessionData ? (
-            <Link href={`/${sessionData?.user.id}`}>
-              <Image
-                src={sessionData?.user.image || ""}
-                alt="profile"
-                width={48}
-                height={18}
-                className="gap-3 rounded-full"
-              />
-            </Link>
-          ) : (
-            <div className="mb-12"></div> //put logo here later
-          )}
-        </div>
-        <button
-          className=" rounded-full font-semibold text-black no-underline transition"
-          onClick={() => setShowModal(true)}
-        >
-          Create
-        </button>
-        <AddPosts
-          showModal={showModal}
-          setShowModal={setShowModal}
-          openModal={openModal}
-          closeModal={closeModal}
-        />
-        <button className=" rounded-full font-semibold text-black no-underline transition">
-          posts
-        </button>
-        <button
-          className={
-            sessionData
-              ? `w-2/12  rounded bg-red-500 font-semibold  text-white hover:bg-red-600`
-              : `w-2/12  rounded bg-blue-500  font-semibold text-white hover:bg-blue-600`
-          }
-          onClick={sessionData ? () => void signOut() : () => void signIn()}
-        >
-          {sessionData ? "Sign out" : "Sign in"}
-        </button>
-      </div>
-    </div>
-  );
-};
+//   return (
+//     <div className="w-full border-slate-400">
+//       <div className="flex gap-5 border-b border-slate-400 p-4">
+//         <div className=" flex w-full gap-3">
+//           {sessionData ? (
+//             <Link href={`/${sessionData?.user.id}`}>
+//               <Image
+//                 src={sessionData?.user.image || ""}
+//                 alt="profile"
+//                 width={48}
+//                 height={18}
+//                 className="gap-3 rounded-full"
+//               />
+//             </Link>
+//           ) : (
+//             <div className="my-5">
+//               {" "}
+//               <Image
+//                 src="/Green-gift-logo.png"
+//                 alt="logo"
+//                 width={160}
+//                 height={100}
+//               ></Image>{" "}
+//             </div> //put logo here later
+//           )}
+//         </div>
+//         <button
+//           className=" rounded-full font-semibold text-black no-underline transition"
+//           onClick={() => setShowModal(true)}
+//         >
+//           Create
+//         </button>
+//         <AddPosts
+//           showModal={showModal}
+//           setShowModal={setShowModal}
+//           openModal={openModal}
+//           closeModal={closeModal}
+//         />
+//         <button className=" rounded-full font-semibold text-black no-underline transition">
+//           posts
+//         </button>
+//         <button
+//           className={
+//             sessionData
+//               ? `w-2/12  rounded bg-red-500 font-semibold  text-white hover:bg-red-600`
+//               : `w-2/12  rounded bg-blue-500  font-semibold text-white hover:bg-blue-600`
+//           }
+//           onClick={sessionData ? () => void signOut() : () => void signIn()}
+//         >
+//           {sessionData ? "Sign out" : "Sign in"}
+//         </button>
+//       </div>
+//     </div>
+//   );
+// };
